Make the spliterator an internal class
